requirements.txt
asgiref==3.8.1
asttokens==2.4.1
black==24.10.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
decorator==5.1.1
defusedxml==0.8.0rc2
Django==5.1.3
django-debug-toolbar==4.4.6
django-extensions==3.2.3
easy-thumbnails==2.10
executing==2.1.0
idna==3.10
ipython==8.29.0
jedi==0.19.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-decouple==3.8
python3-openid==3.2.0
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
